+++ app/src/main/java/org/groundplatform/android/ui/common/BindingAdapters.kt
@@ -25,6 +25,7 @@ import androidx.databinding.BindingAdapter
 import com.google.android.gms.common.SignInButton
 import com.squareup.picasso.Picasso
 import org.groundplatform.android.R
+import org.groundplatform.android.ui.util.RotateTransformation
 import timber.log.Timber
 
 /**
@@ -54,7 +55,7 @@ object BindingAdapters {
       Picasso.get()
         .load(uri)
         .placeholder(R.drawable.ic_photo_grey_600_24dp)
-        .transform(RotateTransformation(uri, view.context)) // Apply transformation
+        .transform(RotateTransformation(uri, view.context))
         .into(view)
     } catch (exception: IllegalStateException) {
       // Needed for running unit tests
